% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_extract_polygons.R
\name{radar_extract_polygons}
\alias{radar_extract_polygons}
\title{Extract mean radar values per polygon}
\usage{
radar_extract_polygons(img, polygons, scale = 10)
}
\arguments{
\item{img}{ee$Image}

\item{polygons}{ee$FeatureCollection}

\item{scale}{Scale in meters (default 10)}
}
\value{
ee$FeatureCollection
}
\description{
Extract mean radar values per polygon
}
