% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_flood.R
\name{radar_flood}
\alias{radar_flood}
\title{Detect flood areas from radar image}
\usage{
radar_flood(img, threshold = -15, aoi = NULL)
}
\arguments{
\item{img}{ee$Image or ee$ImageCollection}

\item{threshold}{Threshold in dB to classify water (default -15)}

\item{aoi}{ee$Geometry (optional)}
}
\value{
ee$Image Mask where water = 1
}
\description{
Detect flood areas from radar image
}
