% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_load.R
\name{radar_load}
\alias{radar_load}
\title{Load Sentinel-1 SAR ImageCollection}
\usage{
radar_load(start_date, end_date, bbox, polarization = "VV")
}
\arguments{
\item{start_date}{Start date (YYYY-MM-DD)}

\item{end_date}{End date (YYYY-MM-DD)}

\item{bbox}{Bounding box as c(xmin, ymin, xmax, ymax)}

\item{polarization}{Polarization type ("VV" or "VH")}
}
\value{
ee$ImageCollection
}
\description{
Load Sentinel-1 SAR ImageCollection
}
