% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_summary.R
\name{radar_summary}
\alias{radar_summary}
\title{Summarize radar ImageCollection}
\usage{
radar_summary(img_col, band = NULL)
}
\arguments{
\item{img_col}{ee$ImageCollection}

\item{band}{Band name to summarize (default: first available band)}
}
\value{
ee$Image
}
\description{
Computes a temporal summary (mean) for a selected band
}
