% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_visualize.R
\name{radar_visualize}
\alias{radar_visualize}
\title{Visualize radar image in Earth Engine Map}
\usage{
radar_visualize(
  img,
  aoi = NULL,
  band = NULL,
  palette = c("black", "blue", "green", "yellow", "red"),
  min = NULL,
  max = NULL
)
}
\description{
Visualize radar image in Earth Engine Map
}
